fix: luksKillSlot (potentially) extra slots/keys when setting LUKS password
the online reencryption creates a new LUKS slot. This breaks the model
of assuming we've only got 1 (primary) or 2 (recovery).

I was seeing the online reencryption happen on the first two boots
(instead of the expecting first boot only).

So now, when changing the LUKS passphrase (e.g. randomizing on the first
boot) loop through all LUKS slots and kill those we don't expect.